refactor(supplier-profile): extract ProductMediaStep and slim orchestrator to compose steps
